requirements.txt
numpy==2.1.1
opencv-python-headless==4.10.0.84
paho-mqtt==2.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
websockets==13.0.1
